+++ include/sora/open_h264_video_encoder.h
@@ -5,12 +5,14 @@
 #include <string>
 
 // WebRTC
+#include <api/environment/environment.h>
 #include <api/video_codecs/sdp_video_format.h>
 #include <api/video_codecs/video_encoder.h>
 
 namespace sora {
 
 std::unique_ptr<webrtc::VideoEncoder> CreateOpenH264VideoEncoder(
+    const webrtc::Environment& env,
     const webrtc::SdpVideoFormat& format,
     std::string openh264);