+++ include/sora/sora_video_decoder_factory.h
@@ -22,6 +22,7 @@ struct VideoDecoderConfig {
   // 指定したコーデックに対応するデコーダを設定する
   VideoDecoderConfig(webrtc::VideoCodecType codec,
                      std::function<std::unique_ptr<webrtc::VideoDecoder>(
+                         const webrtc::Environment&,
                          const webrtc::SdpVideoFormat&)> create_video_decoder)
       : codec(codec), create_video_decoder(std::move(create_video_decoder)) {}
   // 特定の SdpVideoFormat に対応するデコーダを設定する
@@ -29,6 +30,7 @@ struct VideoDecoderConfig {
   VideoDecoderConfig(std::function<std::vector<webrtc::SdpVideoFormat>()>
                          get_supported_formats,
                      std::function<std::unique_ptr<webrtc::VideoDecoder>(
+                         const webrtc::Environment&,
                          const webrtc::SdpVideoFormat&)> create_video_decoder)
       : get_supported_formats(std::move(get_supported_formats)),
         create_video_decoder(std::move(create_video_decoder)) {}
@@ -43,6 +45,7 @@ struct VideoDecoderConfig {
   webrtc::VideoCodecType codec = webrtc::VideoCodecType::kVideoCodecGeneric;
   std::function<std::vector<webrtc::SdpVideoFormat>()> get_supported_formats;
   std::function<std::unique_ptr<webrtc::VideoDecoder>(
+      const webrtc::Environment&,
       const webrtc::SdpVideoFormat&)>
       create_video_decoder;
   std::shared_ptr<webrtc::VideoDecoderFactory> factory;